refactor: Move SeamProvider into seam
