+++ src/lib/index.ts
@@ -1,3 +1,2 @@
 export * from './seam/index.js'
-export * from './SeamProvider.js'
 export * from './ui/index.js'
