+++ src/lib/seam/index.ts
@@ -2,4 +2,5 @@ export * from './access-codes/use-access-code.js'
 export * from './access-codes/use-access-codes.js'
 export * from './devices/use-device.js'
 export * from './devices/use-devices.js'
+export * from './SeamProvider.js'
 export * from './use-seam-client.js'
